Fix mnemonics off by one error (#438)
* Simplify mnemonics implementation

* ebin

* Fix build

* Fix off by one error, fix type coercion

+++ src/zedwallet/Transfer.cpp
@@ -558,7 +558,7 @@ void doTransfer(std::string address, uint64_t amount, uint64_t fee,
     }
 
     p.fee = fee;
-    p.mixIn = mixin;
+    p.mixIn = static_cast<uint16_t>(mixin);
     p.extra = extra;
     p.changeDestination = walletInfo->walletAddress;
